Rename BNA routes
Renames the BNAs routes to Ratings route to reflect to definitions from
the team's glossary.
