+++ lambdas/src/cities/get-cities-ratings.rs
