+++ lambdas/src/ratings/get-ratings-results.rs
