+++ lambdas/src/ratings/mod.rs
