+++ lambdas/src/ratings/patch-ratings.rs
